Fix: test counter increase

+++ counter-native/src/state.rs
@@ -2,5 +2,5 @@ use borsh::{BorshSerialize, BorshDeserialize};
 
 #[derive(BorshSerialize, BorshDeserialize, PartialEq, Debug, Clone)]
 pub struct Counter {
-    pub count: u32
+    pub count: u64
 }
